Added comments

+++ PowerofTwo.cpp
@@ -1,17 +1,21 @@
 #include<iostream>
 using namespace std;
 
-int power(long n){
+int power(int n){
+    // base case
     if(n==0)
     return 1;
-
+    
+    // recursive relation
     return 2*power(n-1);
 }
 
 int main(){
-    long n;
+    int n;
     cin>>n;
-    long ans = power(n);
+
+    // calling function
+    int ans = power(n);
     cout<<ans<<endl;
     return 0;
 }
